Increase accelerator search radius by 1
Having it set at +/- 2 was causing some generated samples that were just
below r to be considered valid.

I don't remember where I did the math to get the +/- 2, so I will have
to recreate that at some point to try and determine what is going on.

Bumping it up to +/- 3 solves this issue.